Move Npm test, delete example test
